refactor(il/verify): remove obsolete monolithic verifier
delete legacy Verifier files after splitting into dedicated classes
